+++ src/test/java/RunnerClass/AcademicsRunner.java
@@ -5,7 +5,7 @@ import cucumber.api.junit.Cucumber;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(features="src\\test\\java\\AcademicsFeaturesFiles",
-                 glue= "AcademicsStepDef")
+@CucumberOptions(features="src\\test\\java\\FutureFiles\\Academics.feature",
+                 glue= "StepDef")
 public class AcademicsRunner {
 }
