+++ src/test/java/RunnerClass/CampusLifeRunner.java
@@ -5,7 +5,7 @@ import cucumber.api.junit.Cucumber;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(features="src\\test\\java\\CampusLifeFeaturesFiles",
-        glue= "CampusLifeStepDef")
+@CucumberOptions(features="src\\test\\java\\FutureFiles\\CampusLife.feature",
+        glue= "StepDef")
 public class CampusLifeRunner {
 }
