+++ src/test/java/StepDef/CampusLifeStepDef.java
@@ -1,4 +1,4 @@
-package CampusLifeStepDef;
+package StepDef;
 
 import Pages.CampusLifePage;
 import cucumber.api.java.en.Given;
@@ -9,38 +9,38 @@ public class CampusLifeStepDef {
     @Given("User will navigate to the NAU website")
     public void user_will_navigate_to_the_NAU_website() {
         // Write code here that turns the phrase above into concrete actions
-        throw new cucumber.api.PendingException();
+
 
     }
 
     @When("User hovers on {string}")
     public void user_hovers_on(String string) {
         // Write code here that turns the phrase above into concrete actions
-        throw new cucumber.api.PendingException();
+
     }
 
     @Then("Three sections on Student Services, Residential Life and Parking & Security are displayed on screen")
     public void three_sections_on_Student_Services_Residential_Life_and_Parking_Security_are_displayed_on_screen() {
         // Write code here that turns the phrase above into concrete actions
-        throw new cucumber.api.PendingException();
+
     }
 
     @Then("Sub-section on Student Services will be highlighted when user hovers over it")
     public void sub_section_on_Student_Services_will_be_highlighted_when_user_hovers_over_it() {
         // Write code here that turns the phrase above into concrete actions
-        throw new cucumber.api.PendingException();
+
     }
 
     @Then("Sub-section on Residential Life will be highlighted when user hovers over it")
     public void sub_section_on_Residential_Life_will_be_highlighted_when_user_hovers_over_it() {
         // Write code here that turns the phrase above into concrete actions
-        throw new cucumber.api.PendingException();
+
     }
 
     @Then("Sub-section on Parking & Security will be highlighted when user hovers over it")
     public void sub_section_on_Parking_Security_will_be_highlighted_when_user_hovers_over_it() {
         // Write code here that turns the phrase above into concrete actions
-        throw new cucumber.api.PendingException();
+
     }
